Implemented validation for backend

+++ mozi-backend/utils/category.ts
@@ -1,4 +1,7 @@
+import { GraphQLError } from "graphql/error";
 import { v4 as uuidv4 } from "uuid";
+import { NOT_VALID_CATEGORY, UNAUTHORIZED_MESSAGE } from "../common/errorMessages";
+import { categorySchema } from "../common/validation";
 import { MyContext } from "../server";
 import { Category } from "./types";
 
@@ -7,9 +10,11 @@ export function getCategories(_: any, context:MyContext):Promise<Category[]> {
   return context.db.all<Category>(sql)
 }
 
-export function getCategoryById(id: string, context:MyContext):Promise<Category|null>{
+export async function getCategoryById(id: string, context:MyContext):Promise<Category|null>{
   const sql = `SELECT * FROM category WHERE category.id = ?`;
-  return context.db.get<Category>(sql,[id])
+  const result = await context.db.get<Category>(sql,[id])
+  if(result === undefined) return null;
+  return result
 }
 
 export async function createCategory(name:string, context:MyContext):Promise<Category> {
@@ -17,27 +22,42 @@ export async function createCategory(name:string, context:MyContext):Promise<Cat
     id:uuidv4(),
     name:name
   }
-  const sql = `INSERT INTO category (id,name) VALUES (?,?)`;
-  context.db.run(sql,[newCategory.id,newCategory.name])
-  return newCategory
+  if(context.user!.role.toString() !== "viewer"){
+    const validation = await categorySchema.isValid(newCategory)
+    if(!validation) throw new GraphQLError(NOT_VALID_CATEGORY,{extensions:{code:'VALIDATION_FAILED'}})
+    const sql = `INSERT INTO category (id,name) VALUES (?,?)`;
+    context.db.run(sql,[newCategory.id,newCategory.name])
+    return newCategory
+  }
+  throw new GraphQLError(UNAUTHORIZED_MESSAGE,{extensions:{code:'UNAUTHORIZED'}})
 }
 
 export async function updateCategory(category: Category, context:MyContext):Promise<Category|null> {
-  const sql = `UPDATE category SET name = ? WHERE category.id = ?`;
-  context.db.run(sql,[category.name,category.id])
-  return getCategoryById(category.id,context)
+  if(context.user!.role.toString() !== "viewer"){
+    const validation = await categorySchema.isValid(category)
+    if(!validation) throw new GraphQLError(NOT_VALID_CATEGORY,{extensions:{code:'VALIDATION_FAILED'}})
+    const sql = `UPDATE category SET name = ? WHERE category.id = ?`;
+    context.db.run(sql,[category.name,category.id])
+    return getCategoryById(category.id,context)
+  }
+  throw new GraphQLError(UNAUTHORIZED_MESSAGE,{extensions:{code:'UNAUTHORIZED'}})
 }
 
 export async function deleteCategory(id: string, context:MyContext):Promise<Category|null> {
-  const sqlDelete = `DELETE FROM category WHERE category.id = ?`;
-  const sqlUpdate = `UPDATE movie SET category_id = "removedID" WHERE category_id = ?` ;
-  const category = getCategoryById(id,context)
-  context.db.run(sqlUpdate,[id])
-  context.db.run(sqlDelete,[id])
-  return category
+  if(context.user!.role.toString() !== "viewer"){
+    const sqlDelete = `DELETE FROM category WHERE category.id = ?`;
+    const sqlUpdate = `UPDATE movie SET category_id = "removedID" WHERE category_id = ?` ;
+    const category = getCategoryById(id,context)
+    context.db.run(sqlUpdate,[id])
+    context.db.run(sqlDelete,[id])
+    return category
+  }
+  throw new GraphQLError(UNAUTHORIZED_MESSAGE,{extensions:{code:'UNAUTHORIZED'}})
 }
 
-export function checkForCategory(name: string, context:MyContext):Promise<Category|null> {
+export async function checkForCategory(name: string, context:MyContext):Promise<Category|null> {
   const sql = `SELECT * FROM category WHERE category.name = ?`;
-  return context.db.get<Category>(sql,[name])
+  const result = await context.db.get<Category>(sql,[name])
+  if(result === undefined) return null;
+  return result
 }
